Add automatic rejection for unread appeal instructions
The code has been updated to introduce an automatic rejection of appeal if a user attempts to submit their appeal in less than 15 seconds, indicating unread instructions. A new database entry will also be generated for these automatic rejections. A note to this effect can be seen in the commit diff above. Additionally, the language in status update notifications has been changed to reflect this new policy.